+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `devctl release create`: Drop `karpenter-nodepools` from AWS releases v32.0.0 and higher.
+- `devctl release create --bumpall`: Improve UX by coloring bumped versions green and removed apps red in the summary table.
+- Add support for dropped apps easily.
+
 ## [7.5.3] - 2025-08-26
 
 ### Fixed